pyproject.toml
[tool.poetry]
name = "github-profile-analyzer"
version = "0.1.0"
description = "AI-driven GitHub profile analyzer for recruiters with insights dashboard"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.0.0"
pygithub = "^2.1.1"
google-generativeai = "^0.3.0"
sqlalchemy = "^2.0.0"
alembic = "^1.12.0"
python-dotenv = "^1.0.0"
httpx = "^0.25.0"
streamlit = "^1.28.0"
pandas = "^2.1.0"
plotly = "^5.17.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
